fix(source): compile on darwin and linux
The workflow's first run failed on macOS at `wails build`, which is the whole
argument for the workflow. internal/source/handle_posix.go is built only on
linux and darwin, so this Windows machine -- the only place the gate has ever
run -- excludes it before the type-checker sees it. unix.FcntlInt takes a
uintptr descriptor and was being handed the int that every neighbouring call
uses. The file has been in the tree since Epic 1, was reviewed, and shipped
inside two epics that were called green.

Converted at the two call sites rather than changing the parameter, which
matches Open, Fstat and Close beside it and the os.NewFile conversion four
lines above. Verified with darwin/arm64, darwin/amd64 and linux/amd64 builds
and a darwin vet before pushing.

AGENTS.md gains that pre-flight, labelled as what it is: a type-check, not
release proof, and no workflow step cross-builds. It also gains a warning that
`gh run watch --exit-status` exited 0 on this very run after printing the
failure it was watching -- trusting that exit code would have closed the story
green over a red macOS job.

+++ internal/source/handle_posix.go
@@ -125,15 +125,20 @@ func (c *posixContent) Close() error {
 	return err
 }
 
+// The descriptor stays an int here because every other unix call in this file
+// takes one -- Open returns it, Fstat and Close consume it. FcntlInt is the
+// lone exception, wanting a uintptr, so the conversion happens at its two call
+// sites rather than changing a parameter that matches all its neighbours. Same
+// conversion os.NewFile needs a few lines above.
 func clearPosixNonBlocking(descriptor int) error {
-	flags, err := unix.FcntlInt(descriptor, unix.F_GETFL, 0)
+	flags, err := unix.FcntlInt(uintptr(descriptor), unix.F_GETFL, 0)
 	if err != nil {
 		return err
 	}
 	if flags&unix.O_NONBLOCK == 0 {
 		return nil
 	}
-	_, err = unix.FcntlInt(descriptor, unix.F_SETFL, flags&^unix.O_NONBLOCK)
+	_, err = unix.FcntlInt(uintptr(descriptor), unix.F_SETFL, flags&^unix.O_NONBLOCK)
 	return err
 }
